requirement.txt
streamlit
pandas
plotly.express
sickit-learn